fix: exporting a constant from within a react tree throws error exports are not initialized before another file is importing that constant within a react component
