+++ src/pages/App.tsx
@@ -19,18 +19,11 @@ import Sidebar from "@/components/sidebar/Sidebar";
 import { uid } from "uid";
 import { INote } from "@/utils/types/Note";
 import SideMenu from "@/components/SideMenu";
-import NavbarActions from "@/components/navbar-actions/NavbarActions";
 import { User, signOut } from "firebase/auth";
-import { FirebaseContext } from "@/App";
+import { FirebaseContext } from "@/utils/contexts/firebaseContext";
 import { placeholders } from "@/utils/placeholderNotes";
-import {
-	DocumentData,
-	QueryDocumentSnapshot,
-	addDoc,
-	collection,
-	doc,
-	setDoc,
-} from "firebase/firestore";
+import { doc, setDoc } from "firebase/firestore";
+import { fetchUserNotesLoader } from "@/loader/fetchUserNotes";
 
 export interface IContextType {
 	notes: Array<INote>;
@@ -50,7 +43,8 @@ export const NavbarActionsContext = createContext<INavbarActions>(
 );
 const App = () => {
 	const navigate = useNavigate();
-	const fetchedNotes = useLoaderData() as QueryDocumentSnapshot[];
+	// const fetchedNotes = useLoaderData() as QueryDocumentSnapshot[];
+	const [fetching, setFetching] = useState(false);
 	const { auth, db } = useContext(FirebaseContext);
 	const noteIDRef = useRef(undefined);
 	const [searchInput, setSearchInput] = useState<string>("");
@@ -83,12 +77,15 @@ const App = () => {
 		[notes]
 	);
 
-	function setLocalStateNotes() {
-		const userNotes = fetchedNotes.map((doc) => {
-			return doc.data();
-		});
+	async function setLocalStateNotes() {
+		// setFetching(true);
+		// const userNotes = (await fetchUserNotesLoader(auth, db)).map((doc) => {
+		// 	console.log(doc.data().dateAdded);
+		// 	return doc.data();
+		// });
+		// setFetching(false);
 		// setNotes((prev) => [userNotes, ...prev]);
-		console.log(userNotes);
+		// console.log(userNotes);
 	}
 
 	async function addNote(): Promise<void> {
