+++ src/utils/types/FirebaseContextProp.d.ts
@@ -1,0 +1,4 @@
+export interface IFirebaseContextProp {
+	auth: Auth;
+	db: Firestore;
+}
